Revert change in 4bf706c0ef6a887c9d312c84d5c216be0c9ffe0f
This reverts the change in 4bf706c0ef6a887c9d312c84d5c216be0c9ffe0f due to it
being updated in the wrong location. Changes need to go into either the
mutt-compile-colors.sh script and/or the
mutt-colors-solarized-template file.

From there, the muttrc files are generated by the shell script.

Issue is that J_base2 is considered "black" on the light terminal,
changing that to J_cyan would place cyan on top of the default
background colour.

Based on some initial testing, I believe it might be a local terminal
colour control problem.